Initial implementation of a "universal" Extractor
