+++ python/light-curve/light_curve/__init__.py
@@ -1,2 +1,7 @@
 from .light_curve_py import *
+
+# Hide Python features with Rust equivalents
 from .light_curve_ext import *
+
+# Hide Rust Extractor with universal Python Extractor
+from .light_curve_py import Extractor
